♻️ DRY  shared styles
