+++ libraries/core-react/src/components/Select/SingleSelect/SingleSelect.tsx
@@ -1,10 +1,5 @@
 import * as React from 'react'
-import {
-  forwardRef,
-  SelectHTMLAttributes,
-  useState,
-  HTMLAttributes,
-} from 'react'
+import { forwardRef, SelectHTMLAttributes, useState } from 'react'
 import {
   useCombobox,
   UseComboboxProps,
@@ -12,15 +7,18 @@ import {
 } from 'downshift'
 import styled from 'styled-components'
 import { Label } from '../../Label'
-import { Button } from '../../Button'
 import { Icon } from '../../Icon'
-import { Input } from '../../TextField/Input'
 import { arrow_drop_down, arrow_drop_up } from '@equinor/eds-icons'
-import { List } from '../../List'
-import { typographyTemplate, spacingsTemplate } from '@utils'
+import { spacingsTemplate } from '@utils'
 import { select as tokens } from '../Select.tokens'
-
-const { ListItem } = List
+import {
+  Container,
+  PaddedInput,
+  StyledList,
+  StyledButton,
+  StyledListItem,
+  StyledInputWrapper,
+} from '../commonStyles'
 
 export type SingleSelectProps = {
   /** List of options to choose from */
@@ -46,59 +44,8 @@ export type SingleSelectProps = {
   handleSelectedItemChange?: (changes: UseComboboxStateChange<string>) => void
 } & SelectHTMLAttributes<HTMLSelectElement>
 
-const StyledInputWrapper = styled.div`
-  position: relative;
-`
-
-const PaddedInput = styled(Input)`
-  /* Hack: Had to add + 0px to satisfy the style lint plugin */
-  padding-right: calc(
-    ${tokens.button.size} + ${tokens.button.spacings.left} +
-      ${tokens.button.spacings.right} + 0px
-  );
-`
-
-const StyledButton = styled(Button)`
-  position: absolute;
-  right: ${tokens.button.spacings.right};
-  top: ${tokens.button.spacings.top};
-  height: ${tokens.button.size};
-  width: ${tokens.button.size};
-`
-
-const StyledList = styled(List)`
-  background-color: ${tokens.background};
-  box-shadow: ${tokens.boxShadow};
-  overflow-y: scroll;
-  max-height: 300px;
-  padding: 0;
-  border-radius: ${tokens.borderRadius};
-  margin-top: 4px;
-  position: absolute;
-  right: 0;
-  left: 0;
-  z-index: 200;
-`
-
-type StyledListItemType = {
-  highlighted: string
-}
-
-const StyledListItem = styled(ListItem)<StyledListItemType>`
-  list-style: none;
-  ${typographyTemplate(tokens.typography)};
+const PaddedStyledListItem = styled(StyledListItem)`
   ${spacingsTemplate(tokens.spacings.single)};
-  margin: 0;
-  background-color: ${({ highlighted }) =>
-    highlighted === 'true' ? tokens.hover.background : tokens.background};
-  cursor: ${({ highlighted }) =>
-    highlighted === 'true' ? 'pointer' : 'default'};
-`
-
-type ContainerProps = HTMLAttributes<HTMLDivElement>
-
-const Container = styled.div<ContainerProps>`
-  position: relative;
 `
 
 export const SingleSelect = forwardRef<HTMLDivElement, SingleSelectProps>(
@@ -177,13 +124,13 @@ export const SingleSelect = forwardRef<HTMLDivElement, SingleSelectProps>(
         <StyledList {...getMenuProps()}>
           {isOpen &&
             inputItems.map((item, index) => (
-              <StyledListItem
+              <PaddedStyledListItem
                 highlighted={highlightedIndex === index ? 'true' : 'false'}
                 key={`${item}`}
                 {...getItemProps({ item, index })}
               >
                 {item}
-              </StyledListItem>
+              </PaddedStyledListItem>
             ))}
         </StyledList>
       </Container>
